Continue file content logic in this commit
Because of the workflow of the addition of the file content
(OCR using Tesseract), files are now shipped within the card form JSON
on card creation, and not separately as before. The app is refactored
to take into account all needed changes. SQL FTS, at least with SQLite,
is now way too complicated. Removing all the logic to switch over to
ElasticSearch, way easier and fitted for that usage.
Create and Search works. Update and Delete not tested.

Front-end :
- Disable auto upload

- Remove "onUpload..." methods

- fileContent field is NOT added at the moment, must go through API for now

Back-end :
- Change workflow to store files, now shipped with card info

- Fix OCR Tesseract logic

- Beautify and refactor SQL queries

- Remove usage of SQLite FTS entirely

- Add another abstract repository for FTS
  -> This will permit to add more FTS technologies in the future

- Add ElasticSearch -> Works perfectly

- Typo and mispelling corrector are now supported